Made CompoundInterest() use decimal like all other financial applications

+++ CainsSharpTools/MathTools.cs
@@ -140,13 +140,13 @@ namespace CainsSharpTools
         /// <param name="annualRateMultiplier">The multiplier for each compound Note: if less than 1, amount will DECREASE.</param>
         /// <param name="timesCompoundedPerYear">How many times the interest compounds each year</param>
         /// <returns>The value after compound interest</returns>
-        public static float CompoundInterest(float startNumber, int numberOfYears, float annualRateMultiplier,
+        public static decimal CompoundInterest(decimal startNumber, int numberOfYears, float annualRateMultiplier,
             int timesCompoundedPerYear = 1)
         {
             var afterAmount = startNumber;
             for (var i = 0; i < numberOfYears; i++)
             {
-                startNumber += (annualRateMultiplier / timesCompoundedPerYear);
+                startNumber += (decimal)(annualRateMultiplier / timesCompoundedPerYear);
             }
 
             return afterAmount;
